requirements.txt
Flask
render_template
request
send_file
redirect
session
url_for
qrcode
os
base64
sqlite3
werkzeug.security
io